libmix: fix some compatibility in decode to support 64-bit user space
BZ: 178142

fix some compatibility(H264/MPEG4/H263) in decode to
support 64-bit user space

+++ mixvbp/vbp_manager/vbp_mp42_parser.c
@@ -681,7 +681,7 @@ uint32 vbp_get_sc_pos_mp42(
     {
         /* Check if we are byte aligned & phase=0, if thats the case we can check
          work at a time instead of byte*/
-        if (((((uint32) ptr) & 0x3) == 0) && (phase == 0))
+        if (((((uintptr_t) ptr) & 0x3) == 0) && (phase == 0))
         {
             while (data_left > 3)
             {
